--- src/components/tree/nodes.js
@@ -1,137 +1,0 @@
-import React from "react";
-import Radium from "radium";
-import moment from "moment";
-import "moment-range";
-// import _ from "lodash";
-// import Flex from "./framework/flex";
-import { connect } from "react-redux";
-// import { FOO } from "../actions";
-import Node from "./node";
-import Tooltip from "./tooltip";
-import {VictoryAnimation} from "victory";
-
-@connect((state) => {
-  return {controls: state.controls};
-})
-@Radium
-class Nodes extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-
-    };
-  }
-  static propTypes = {
-    /* react */
-    // dispatch: React.PropTypes.func,
-    params: React.PropTypes.object,
-    routes: React.PropTypes.array,
-    /* component api */
-    style: React.PropTypes.object
-    // foo: React.PropTypes.string
-  }
-
-  xVal(node, distanceMeasure, layout) {
-    return this.props.xScale(node.geometry[distanceMeasure][layout].xVal);
-  }
-  yVal(node, distanceMeasure, layout) {
-    return this.props.yScale(node.geometry[distanceMeasure][layout].yVal);
-  }
-  xMidpoint(node, distanceMeasure, layout) {
-    return this.props.xScale(node.geometry[distanceMeasure][layout].xValMidpoint);
-  }
-  yMidpoint(node, distanceMeasure, layout) {
-    return this.props.yScale(node.geometry[distanceMeasure][layout].yValMidpoint);
-  }
-  r_x(node, distanceMeasure, layout) {
-    if (layout === "radial") {
-      return this.props.xScale(node.geometry[distanceMeasure][layout].radiusInner) - this.props.xScale(0);
-    } else {
-      return 0;
-    }
-  }
-  r_y(node, distanceMeasure, layout) {
-    if (layout === "radial") {
-      return this.props.yScale(node.geometry[distanceMeasure][layout].radiusInner) - this.props.yScale(0);
-    } else {
-      return 0;
-    }
-  }
-  smallBigArc(node, distanceMeasure, layout) {
-    if (layout === "radial") {
-      return node.geometry[distanceMeasure][layout].smallBigArc;
-    } else {
-      return 0;
-    }
-  }
-  leftRight(node, distanceMeasure, layout) {
-    if (layout === "radial") {
-      return node.geometry[distanceMeasure][layout].leftRight;
-    } else {
-      return 0;
-    }
-  }
-
-  drawNodes(nodes) {
-    const range = moment().range(
-      new Date(+this.props.query.dmin),
-      new Date(+this.props.query.dmax)
-    )
-    const nodeComponents = nodes.map((node, index) => {
-      return (
-        <VictoryAnimation duration={1000} key={index} data={{
-          x: this.xVal(node, this.props.distanceMeasure, this.props.layout),
-          y: this.yVal(node, this.props.distanceMeasure, this.props.layout),
-          midpoint_x: this.xMidpoint(node, this.props.distanceMeasure, this.props.layout),
-          midpoint_y: this.yMidpoint(node, this.props.distanceMeasure, this.props.layout),
-          source_x:   this.xVal(node.parent, this.props.distanceMeasure, this.props.layout),
-          source_y:   this.yVal(node.parent, this.props.distanceMeasure, this.props.layout),
-          r_x: this.r_x(node, this.props.distanceMeasure, this.props.layout),
-          r_y: this.r_y(node, this.props.distanceMeasure, this.props.layout),
-          smallBigArc: this.smallBigArc(node, this.props.distanceMeasure, this.props.layout),
-          leftRight: this.leftRight(node, this.props.distanceMeasure, this.props.layout)
-        }}>
-        {(props) => {
-          return (
-            <Node
-              {...this.props} {...props} animate={null}
-              key={index}
-              node={node}
-              dateRange={range}
-              fill={this.props.nodeColor ? this.props.nodeColor(node) : "CCC"}
-              showBranchLabels={this.props.controls.showBranchLabels}
-              strain={node.attr.strain}
-              hasChildren={node.children ? true : false}
-              tipRadius={this.props.tipRadius ? this.props.tipRadius(node) : (node.children ? 0 : 4) }
-              branchStrokeColor={this.props.branchStrokeColor ? this.props.branchStrokeColor(node) : "CCC"}
-              branchStrokeWidth={this.props.branchStrokeWidth ? this.props.branchStrokeWidth(node) : 2}
-              nodeStyle={this.props.nodeStyle ? this.props.nodeStyle(node) : "circle"}
-            />
-          )
-        }}
-      </VictoryAnimation>
-     );
-    });
-    return nodeComponents;
-  }
-
-  drawTooltip(node, type) {
-    return (
-      <Tooltip
-        type={type}
-        node={node}
-        x={this.xVal(node, this.state.distanceMeasure, this.props.layout)}
-        y={this.yVal(node, this.state.distanceMeasure, this.props.layout)}
-      />
-    )
-  }
-  render() {
-    return (
-      <g>
-        {this.drawNodes(this.props.nodes)}
-      </g>
-    );
-  }
-}
-
-export default Nodes;
